+++ src/Map.ts
@@ -17,8 +17,6 @@ import type {
   FilterSpecification,
   StyleSetterOptions,
   ExpressionSpecification,
-  RequestParameters,
-  GetResourceResponse,
 } from "maplibre-gl";
 import { ReferenceMapStyle, MapStyleVariant } from "@maptiler/client";
 import { config, MAPTILER_SESSION_ID, SdkConfig } from "./config";
@@ -42,7 +40,7 @@ import { FullscreenControl } from "./MLAdapters/FullscreenControl";
 
 import Minimap from "./Minimap";
 import type { MinimapOptionsInput } from "./Minimap";
-import { addProtocol } from ".";
+import { registerLocalCacheProtocol } from "./caching";
 
 export type LoadWithTerrainEvent = {
   type: "loadWithTerrain";
@@ -215,88 +213,7 @@ export class Map extends maplibregl.Map {
     });
 
     if (config.caching) {
-      // TODO cleanup
-      // TODO make safer
-      // TODO cache cleaning ?
-      // TODO move elsewhere ?
-      addProtocol(
-        "localcache",
-        async (
-          params: RequestParameters,
-          abortController: AbortController,
-          // eslint-disable-next-line @typescript-eslint/no-explicit-any
-        ): Promise<GetResourceResponse<any>> => {
-          if (!params.url) throw new Error("");
-
-          params.url = params.url.replace("localcache://", "https://");
-
-          if (params.url.includes("tiles.json")) {
-            const requestInit: RequestInit = params;
-            requestInit.signal = abortController.signal;
-            const response = await fetch(params.url, requestInit);
-            const json = await response.json();
-
-            // move `Last-Modified` to query so it propagates to tile URLs
-            json.tiles[0] +=
-              "&last-modified=" + response.headers.get("Last-Modified");
-
-            return {
-              data: json,
-              cacheControl: response.headers.get("Cache-Control"),
-              expires: response.headers.get("Expires"),
-            };
-          } else {
-            const url = new URL(params.url);
-
-            const cacheableUrl = new URL(url);
-            cacheableUrl.searchParams.delete("mtsid");
-            cacheableUrl.searchParams.delete("key");
-            const cacheKey = cacheableUrl.toString();
-
-            const fetchableUrl = new URL(url);
-            fetchableUrl.searchParams.delete("last-modified");
-            const fetchUrl = fetchableUrl.toString();
-
-            const respond = async (
-              response: Response,
-              // eslint-disable-next-line @typescript-eslint/no-explicit-any
-            ): Promise<GetResourceResponse<any>> => {
-              const parsePromise =
-                params.type === "arrayBuffer" || params.type === "image"
-                  ? response.arrayBuffer()
-                  : params.type === "json"
-                  ? response.json()
-                  : response.text();
-              const data = await parsePromise;
-
-              return {
-                data,
-                cacheControl: response.headers.get("Cache-Control"),
-                expires: response.headers.get("Expires"),
-              };
-            };
-
-            // TODO configurable cache name ?
-            const cache = await caches.open("maptiler_tiles");
-
-            const cacheMatch = await cache.match(cacheKey);
-
-            if (cacheMatch) {
-              console.log("Cache hit", cacheKey);
-              return respond(cacheMatch);
-            } else {
-              console.log("Cache miss", cacheKey);
-              const requestInit: RequestInit = params;
-              requestInit.signal = abortController.signal;
-              const response = await fetch(fetchUrl, requestInit);
-              if (response.status < 300) {
-                cache.put(cacheKey, response.clone());
-              }
-              return respond(response);
-            }
-          }
-        },
-      );
+      registerLocalCacheProtocol();
     }
 
     this.primaryLanguage = options.language ?? config.primaryLanguage;
